refactor(installer): consolidate installer checks
